stability fixes

+++ rapp.h
@@ -74,7 +74,7 @@ public:
 	VOID CheckForUpdates (BOOL is_periodical);
 #endif // _APP_NO_UPDATES
 
-	DWORD ConfigGet (LPCWSTR key, INT def, LPCWSTR name = nullptr);
+	LONGLONG ConfigGet (LPCWSTR key, INT def, LPCWSTR name = nullptr);
 	rstring ConfigGet (LPCWSTR key, LPCWSTR def, LPCWSTR name = nullptr);
 
 	BOOL ConfigSet (LPCWSTR key, LONGLONG val, LPCWSTR name = nullptr);
